pyproject.toml
[project]
name = "telegram-obsidian-capture"
version = "0.1.1"
description = "Telegram bot for quick capture to Obsidian vault"
readme = "README.md"
license = "LicenseRef-Polyform-Noncommercial-1.0.0"
authors = [{ name = "Matteo Cervelli", email = "matteo@adli.men" }]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "pydub>=0.25",
    "structlog>=24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
